refactor(external-api): decompose ExternalApiScheduler into 3 phase classes
Mechanical extraction of god-class (495→120 lines) into:
- OcidLookupPhase: OCID lookup via FetchExternalApiUseCase
- CharacterBasicSnapshotPhase: batch snapshot with ChunkedSnapshotSink
- ItemEquipmentSnapshotPhase: continuous loop snapshot
- SchedulerPhaseUtils: shared rate limiter, logging, run ID utilities

ExternalApiScheduler retains only: scheduling, lock management, OCID cache
lifecycle, executor lifecycle, and startup/shutdown orchestration.

No behavior changes. All metrics, Kafka events, snapshot formats, and
runId/chunkId/objectKey naming preserved identically.
